@@ -3,66 +3,55 @@
   "version": "4.0.3-0",
   "description": "Minimal encryption/decryption JWE/CWE library.",
   "license": "BSD-3-Clause",
-  "main": "index.js",
-  "scripts": {
-    "test": "cd test && npm run test-node && karma start karma.conf.js",
-    "test-node": "cd test && npm run test-node",
-    "test-karma": "cd test && karma start karma.conf.js",
-    "coverage": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text-summary npm run test-node",
-    "coverage-ci": "cross-env NODE_ENV=test nyc --reporter=lcovonly npm run test-node",
-    "coverage-report": "nyc report",
-    "lint": "eslint ."
+  "type": "module",
+  "exports": "./lib/index.js",
+  "browser": {
+    "./lib/algorithms/x25519-helper.js": "./lib/algorithms/x25519-helper-browser.js",
+    "./lib/crypto.js": "./lib/crypto-browser.js",
+    "./lib/util.js": "./lib/util-browser.js"
   },
   "files": [
-    "Cipher.js",
-    "DecryptTransformer.js",
-    "EncryptTransformer.js",
-    "algorithms/*.js",
-    "crypto-browser.js",
-    "crypto.js",
-    "index.js",
-    "main.js",
-    "util-browser.js",
-    "util.js"
+    "lib"
   ],
+  "scripts": {
+    "test": "npm run lint && cd test && npm run test-node && npm run test-karma",
+    "test-node": "cross-env NODE_ENV=test mocha -t 30000 -R ${REPORTER:-spec} test/unit/index.js",
+    "test-karma": "cd test && karma start karma.conf.cjs",
+    "coverage": "cross-env NODE_ENV=test c8 --reporter=lcov --reporter=text-summary npm run test-node",
+    "coverage-ci": "cross-env NODE_ENV=test c8 --reporter=lcovonly --reporter=text npm run test-node",
+    "coverage-report": "c8 report",
+    "lint": "eslint ."
+  },
   "dependencies": {
-    "@stablelib/chacha20poly1305": "^1.0.0",
-    "@stablelib/xchacha20poly1305": "^1.0.0",
-    "base58-universal": "^1.0.0",
-    "base64url-universal": "^1.1.0",
-    "esm": "^3.2.25",
-    "isomorphic-webcrypto": "^2.3.2",
-    "tweetnacl": "^1.0.1",
+    "@stablelib/chacha20poly1305": "^1.0.1",
+    "@stablelib/xchacha20poly1305": "^1.0.1",
+    "base58-universal": "^2.0.0",
+    "base64url-universal": "^2.0.0",
+    "isomorphic-webcrypto": "^2.3.8",
+    "tweetnacl": "^1.0.3",
     "web-streams-polyfill": "~3.0.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.13.10",
-    "@babel/plugin-transform-modules-commonjs": "^7.13.8",
-    "@babel/plugin-transform-runtime": "^7.13.10",
-    "@babel/preset-env": "^7.13.10",
-    "@babel/runtime": "^7.13.10",
-    "@digitalbazaar/did-io": "^1.0.0",
-    "@digitalbazaar/did-method-key": "^2.0.0",
-    "@digitalbazaar/x25519-key-agreement-key-2020": "^2.0.0",
-    "babel-loader": "^8.2.2",
-    "chai": "^4.3.3",
-    "chai-bytes": "^0.1.2",
+    "@digitalbazaar/did-io": "^2.0.0",
+    "@digitalbazaar/did-method-key": "^3.0.0",
+    "@digitalbazaar/x25519-key-agreement-key-2020": "^3.0.0",
+    "c8": "^7.11.2",
+    "chai": "^4.3.6",
     "cross-env": "^7.0.3",
-    "eslint": "^7.21.0",
-    "eslint-config-digitalbazaar": "^2.6.1",
-    "eslint-plugin-jsdoc": "^32.2.0",
-    "karma": "^6.2.0",
-    "karma-babel-preprocessor": "^8.0.0",
+    "eslint": "^8.14.0",
+    "eslint-config-digitalbazaar": "^3.0.0",
+    "eslint-plugin-jsdoc": "^39.2.9",
+    "eslint-plugin-unicorn": "^42.0.0",
+    "karma": "^6.3.19",
     "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^3.1.0",
+    "karma-chrome-launcher": "^3.1.1",
     "karma-mocha": "^2.0.1",
     "karma-mocha-reporter": "^2.2.5",
     "karma-sourcemap-loader": "^0.3.8",
     "karma-webpack": "^5.0.0",
-    "mocha": "^8.3.1",
+    "mocha": "^9.2.2",
     "mocha-lcov-reporter": "^1.3.0",
-    "nyc": "^15.1.0",
-    "webpack": "^4.46.0"
+    "webpack": "^5.72.0"
   },
   "repository": {
     "type": "git",
@@ -87,13 +76,12 @@
     "url": "https://github.com/digitalbazaar/minimal-cipher/issues"
   },
   "homepage": "https://github.com/digitalbazaar/minimal-cipher",
-  "module": "main.js",
-  "browser": {
-    "./algorithms/x25519-helper.js": "./algorithms/x25519-helper-browser.js",
-    "./crypto.js": "./crypto-browser.js",
-    "./util.js": "./util-browser.js"
+  "c8": {
+    "exclude": [
+      "test"
+    ]
   },
   "engines": {
-    "node": ">=12.0.0"
+    "node": ">=14"
   }
 }
