@@ -1,17 +1,15 @@
 /*!
- * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2021-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const {CachedResolver} = require('@digitalbazaar/did-io');
-const {driver: didKeyDriver} = require('@digitalbazaar/did-method-key');
+import {CachedResolver} from '@digitalbazaar/did-io';
+import {driver as didKeyDriver} from '@digitalbazaar/did-method-key';
 
 const resolver = new CachedResolver();
 
 // Config did-io to support did:key driver
 resolver.use(didKeyDriver());
 
-function createKeyResolver() {
+export function createKeyResolver() {
   return async function keyResolver({id} = {}) {
     if(!id.startsWith('did:')) {
       throw new Error(`Key ID "${id}" not supported in resolver.`);
@@ -19,5 +17,3 @@ function createKeyResolver() {
     return resolver.get({did: id});
   };
 }
-
-module.exports.createKeyResolver = createKeyResolver;
