@@ -1,8 +1,8 @@
 /*!
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import base64url from 'base64url-universal';
-import {TextDecoder, TransformStream, stringToUint8Array} from './util.js';
+import * as base64url from 'base64url-universal';
+import {TransformStream, stringToUint8Array} from './util.js';
 import {DecryptTransformer} from './DecryptTransformer.js';
 import {EncryptTransformer} from './EncryptTransformer.js';
 import * as fipsAlgorithm from './algorithms/fips.js';
