@@ -1,10 +1,10 @@
 /*!
- * Copyright (c) 2019-2020 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import base64url from 'base64url-universal';
+import * as base64url from 'base64url-universal';
 import * as fipsAlgorithm from './algorithms/fips.js';
 import * as recAlgorithm from './algorithms/recommended.js';
-import {TextDecoder, stringToUint8Array} from './util.js';
+import {stringToUint8Array} from './util.js';
 
 // support `C20P` for backwards compatibility
 import * as c20p from './algorithms/c20p.js';
