@@ -1,7 +1,7 @@
 /*!
- * Copyright (c) 2019-2020 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import base64url from 'base64url-universal';
+import * as base64url from 'base64url-universal';
 
 // 1 MiB = 1048576
 const DEFAULT_CHUNK_SIZE = 1048576;
