@@ -1,11 +1,10 @@
 /*!
- * Copyright (c) 2019-2020 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-import base64url from 'base64url-universal';
+import * as base64url from 'base64url-universal';
 import {createKek} from './aeskw.js';
 import * as base58btc from 'base58-universal';
 import {deriveKey} from './ecdhkdf.js';
-import {TextEncoder} from '../util.js';
 import {deriveSecret, deriveEphemeralKeyPair} from './x25519-helper.js';
 
 const KEY_TYPE = 'X25519KeyAgreementKey2020';
