@@ -1,14 +1,11 @@
-// browser TextDecoder/TextEncoder
 /* eslint-env browser */
-const TextDecoder = self.TextDecoder;
-const TextEncoder = self.TextEncoder;
-export {TextDecoder, TextEncoder};
-
 let ReadableStream = self.ReadableStream;
 let TransformStream = self.TransformStream;
 // TODO: ensure this polyfill is necessary
-import {ReadableStream as Readable, TransformStream as Transform}
-  from 'web-streams-polyfill/ponyfill';
+import {
+  ReadableStream as Readable,
+  TransformStream as Transform
+} from 'web-streams-polyfill/dist/ponyfill.mjs';
 if(!ReadableStream) {
   ReadableStream = Readable;
 }
