@@ -1,8 +1,6 @@
-// Node.js TextDecoder/TextEncoder
-import {TextDecoder, TextEncoder} from 'util';
-export {TextDecoder, TextEncoder};
-
-export {ReadableStream, TransformStream} from 'web-streams-polyfill/ponyfill';
+export {
+  ReadableStream, TransformStream
+} from 'web-streams-polyfill/dist/ponyfill.mjs';
 
 export function stringToUint8Array(data) {
   if(typeof data === 'string') {
